+++ src/components/ProjectBrowser/ProjectBrowser.tsx
@@ -1,36 +1,131 @@
 import * as Tabs from "@radix-ui/react-tabs";
 import { demos } from "../../data/demos";
+import type { Example, ExamplesDataRecord } from "../../data/demos";
 import { ProjectEntry } from "./ProjectEntry";
 import "./ProjectBrowser.css";
 import { assets } from "@kaplayjs/crew";
-import { useCallback, useState } from "react";
+import { Fragment, useCallback, useState } from "react";
 import { Tooltip } from "react-tooltip";
+import examplesJson from "../../../kaplay/examples/examples.json";
 import { useProject } from "../../hooks/useProject";
 import type { ProjectMode } from "../../stores/project";
+import { cn } from "../../util/cn";
 import { TabsList } from "../UI/TabsList";
 import { TabTrigger } from "../UI/TabTrigger";
+import { GroupBy, groupBy } from "./GroupBy";
 import { ProjectCreate } from "./ProjectCreate";
+import {
+    SortBy,
+    sortEntries,
+    sortMapExamples,
+    sortMapProjects,
+} from "./SortBy";
+import { TagsFilter } from "./TagsFilter";
+
+export type ExamplesData = {
+    categories?: ExamplesDataRecord;
+    tags?: ExamplesDataRecord;
+    difficulties?: { displayName?: string }[];
+};
 
 export const ProjectBrowser = () => {
     const [tab, setTab] = useState("Projects");
-    const { getSavedProjects } = useProject();
+    const { getSavedProjects, getProjectMetadata } = useProject();
     const [filter, setFilter] = useState("");
-    const filteredExamples = useCallback(
-        () =>
-            demos.filter((example) =>
-                example.formattedName
-                    .toLowerCase()
-                    .includes(filter.toLowerCase())
-            ),
-        [filter],
+    const [filterTags, setFilterTags] = useState<string[]>([]);
+    const [projectsSort, setProjectsSort] = useState("latest");
+    const [examplesSort, setExamplesSort] = useState("topic");
+    const [projectsGroup, setProjectsGroup] = useState("type");
+    const [examplesGroup, setExamplesGroup] = useState("category");
+
+    const examplesData = examplesJson as ExamplesData;
+
+    const currentSort = useCallback(
+        () => tab == "Projects" ? projectsSort : examplesSort,
+        [tab, projectsSort, examplesSort],
     );
+    const setCurrentSort = (value: string) => {
+        if (tab == "Projects") {
+            setProjectsSort(value);
+        } else {
+            setExamplesSort(value);
+        }
+    };
+    const sortFn = (a: Example, b: Example): number =>
+        sortEntries(currentSort(), tab, a, b);
+
+    const projects = useCallback(
+        () => getSavedProjects().map(project => getProjectMetadata(project)),
+        [getSavedProjects],
+    );
+
+    const currentGroup = useCallback(
+        () => tab == "Projects" ? projectsGroup : examplesGroup,
+        [tab, projectsGroup, examplesGroup],
+    );
+    const setCurrentGroup = (value: string) => {
+        if (tab == "Projects") {
+            setProjectsGroup(value);
+        } else {
+            setExamplesGroup(value);
+        }
+    };
+
     const filteredProjects = useCallback(
         () =>
-            getSavedProjects().filter((project) =>
-                project.toLowerCase().includes(filter.toLowerCase())
+            groupBy(
+                (projects() as Example[]).filter((project) =>
+                    project.name.toLowerCase().includes(filter.toLowerCase())
+                    && (!filterTags.length
+                        || project.tags.some(({ name }) =>
+                            filterTags.includes(name)
+                        ))
+                ).sort(sortFn),
+                projectsGroup,
             ),
-        [filter],
+        [filter, filterTags, currentSort, projectsGroup],
     );
+    const filteredExamples = useCallback(
+        () =>
+            groupBy(
+                demos.filter((example) =>
+                    example.formattedName
+                        .toLowerCase()
+                        .includes(filter.toLowerCase())
+                    && (!filterTags.length
+                        || example.tags.some(({ name }) =>
+                            filterTags.includes(name)
+                        ))
+                ).sort(sortFn),
+                examplesGroup,
+            ),
+        [filter, filterTags, currentSort, examplesGroup],
+    );
+
+    const tags = useCallback((): string[] => {
+        const set = new Set<string>();
+        ((tab == "Projects" ? projects() : demos) as Example[]).sort(
+            sortFn,
+        ).forEach(
+            entry => {
+                if (typeof entry !== "string") {
+                    entry?.tags?.forEach(({ name }) => set.add(name));
+                }
+            },
+        );
+
+        return [...set];
+    }, [tab]);
+    const toggleTag = (tag: string) =>
+        setFilterTags(prev =>
+            prev.includes(tag)
+                ? prev.filter(t => t !== tag)
+                : [...prev, tag]
+        );
+
+    const entriesCount = (entries: object): number =>
+        Object.values(entries).reduce((sum, items) => sum + items.length, 0);
+
     const createButtons = Object.entries({
         pj: "With tree structure",
         ex: "Single file script",
@@ -44,17 +139,55 @@ export const ProjectBrowser = () => {
 
     return (
         <dialog id="examples-browser" className="modal bg-[#00000070]">
-            <div className="modal-box | overflow-hidden max-w-screen-md p-0 flex flex-col w-dvw h-dvh">
-                <header className="grow-0 space-y-4 bg-base-200 p-4 border border-b-0 border-px border-base-100 rounded-t-2xl">
-                    <h2 className="text-3xl text-white font-semibold">
-                        Projects Browser
-                    </h2>
+            <div className="modal-box overflow-hidden max-w-screen-md p-0 flex flex-col w-dvw h-dvh max-h-[calc(100dvh-4.4rem)]">
+                <header
+                    className="grow-0 space-y-4 bg-base-200 p-4 border border-b-0 border-px border-base-100 rounded-t-2xl"
+                    tabIndex={0}
+                >
+                    <div className="flex flex-wrap items-baseline justify-between gap-3">
+                        <h2 className="shrink-0 text-3xl text-white font-semibold">
+                            Projects Browser
+                        </h2>
+
+                        <div className="flex gap-3">
+                            <GroupBy
+                                value={currentGroup()}
+                                onChange={setCurrentGroup}
+                                options={tab == "Projects"
+                                    ? ["type", "none"]
+                                    : [
+                                        "category",
+                                        "group",
+                                        "difficulty",
+                                        "none",
+                                    ]}
+                            />
+                            <SortBy
+                                value={currentSort()}
+                                onChange={setCurrentSort}
+                                options={Object.keys(
+                                    tab == "Projects"
+                                        ? sortMapProjects
+                                        : sortMapExamples,
+                                )}
+                            />
+                        </div>
+                    </div>
 
                     <input
                         type="search"
                         placeholder="Search for examples/projects"
-                        className="input | input-bordered | w-full"
+                        className="input input-bordered w-full"
                         onChange={(ev) => setFilter(ev.target.value)}
+                    />
+
+                    <TagsFilter
+                        value={tags}
+                        tags={examplesData.tags}
+                        filterTags={filterTags}
+                        setFilterTags={setFilterTags}
+                        multipleTags={tab != "Projects"}
+                        tagsExpandingDeps={[tab]}
                     />
                 </header>
 
@@ -68,51 +201,104 @@ export const ProjectBrowser = () => {
                             label="My Projects & Examples"
                             value="Projects"
                             icon={assets.api_book.outlined}
-                            count={filteredProjects().length}
+                            count={entriesCount(filteredProjects())}
                         />
                         <TabTrigger
                             label="KAPLAY Demos"
                             value="Examples"
                             icon={assets.apple.outlined}
-                            count={filteredExamples().length}
+                            count={entriesCount(filteredExamples())}
                         />
                     </TabsList>
 
                     <Tabs.Content
                         value="Projects"
-                        className="flex-1 p-5 pt-4 overflow-auto scrollbar-thin"
+                        className="flex-1 px-4 pb-5 pt-4 overflow-auto scrollbar-thin"
                         tabIndex={-1}
+                        forceMount
+                        hidden={tab !== "Projects"}
                     >
                         {getSavedProjects().length > 0
                             ? (
-                                <div className="examples-list gap-2 rounded-lg">
-                                    {filteredProjects().length > 0
-                                        && filteredProjects().map((project) => (
-                                            <ProjectEntry
-                                                project={{
-                                                    formattedName: project
-                                                        .slice(
-                                                            3,
-                                                        ),
-                                                    name: project,
-                                                    tags: [
-                                                        ...project.startsWith(
-                                                                "pj-",
-                                                            )
-                                                            ? ["project"]
-                                                            : ["example"],
-                                                    ],
-                                                    description: "",
-                                                    id: 0,
-                                                    version: "2.0.0",
-                                                    difficulty: "",
-                                                }}
-                                                isProject
-                                                key={project}
-                                            />
-                                        ))}
-                                    {createButtons}
-                                </div>
+                                <>
+                                    {Object.entries(filteredProjects()).map((
+                                        [groupName, projects],
+                                        index,
+                                    ) => (
+                                        <Fragment key={index}>
+                                            <div
+                                                className={cn(
+                                                    "rounded-lg collapse collapse-arrow px-1 first:pt-0.5",
+                                                    {
+                                                        "first:-mt-3":
+                                                            groupName != "all",
+                                                    },
+                                                )}
+                                            >
+                                                <input
+                                                    className="peer min-h-12"
+                                                    type="checkbox"
+                                                    defaultChecked
+                                                    hidden={groupName == "all"}
+                                                />
+
+                                                {groupName != "all" && (
+                                                    <>
+                                                        <div
+                                                            className={cn(
+                                                                "collapse-title flex items-center gap-1.5 font-medium capitalize px-0 py-3.5 min-h-12 after:!top-7 after:!right-2 peer-hover:text-white transition-colors",
+                                                                {
+                                                                    "text-base-content/50":
+                                                                        groupName
+                                                                            == "uncategorized",
+                                                                },
+                                                            )}
+                                                        >
+                                                            <span className="badge badge-xs font-bold text-[inherit] text-[0.625rem] py-1 px-1 min-w-5 h-auto bg-base-content/15 border-0">
+                                                                {projects
+                                                                    .length}
+                                                            </span>
+                                                            {examplesData
+                                                                .categories
+                                                                ?.[groupName]
+                                                                ?.displayName
+                                                                ?? groupName}
+                                                        </div>
+                                                    </>
+                                                )}
+
+                                                <div
+                                                    className={cn(
+                                                        "examples-list gap-2 collapse-content -mx-0.5 !p-0 peer-checked:!pb-5",
+                                                        {
+                                                            "-mt-5 peer-checked:!pt-5":
+                                                                groupName
+                                                                    != "all",
+                                                        },
+                                                    )}
+                                                >
+                                                    {projects.map((project) => (
+                                                        <ProjectEntry
+                                                            project={project}
+                                                            isProject
+                                                            key={project.name}
+                                                            toggleTag={toggleTag}
+                                                        />
+                                                    ))}
+                                                </div>
+                                            </div>
+                                            <div className="mx-1 border-b border-base-content/10">
+                                            </div>
+                                        </Fragment>
+                                    ))}
+
+                                    <div
+                                        key={"create-button"}
+                                        className="sticky -bottom-5 examples-list gap-2 -mb-5 pt-5 pb-5 px-0.5 bg-base-100"
+                                    >
+                                        {createButtons}
+                                    </div>
+                                </>
                             )
                             : (
                                 <div className="flex flex-col items-center justify-center gap-3 w-full max-w-sm mx-auto min-h-full pb-[4vh]">
@@ -153,9 +339,11 @@ export const ProjectBrowser = () => {
 
                                         Check out KAPLAY Demos
 
-                                        {!!filteredExamples().length && (
+                                        {!!entriesCount(filteredExamples()) && (
                                             <span className="badge badge-xs font-medium py-1 px-1.5 min-w-5 h-auto bg-base-content/15 border-0">
-                                                {filteredExamples().length}
+                                                {entriesCount(
+                                                    filteredExamples(),
+                                                )}
                                             </span>
                                         )}
                                     </button>
@@ -164,14 +352,92 @@ export const ProjectBrowser = () => {
                     </Tabs.Content>
                     <Tabs.Content
                         value="Examples"
-                        className="p-5 pt-4 overflow-auto scrollbar-thin"
+                        className="flex-1 px-4 pb-5 pt-4 overflow-auto scrollbar-thin"
                         tabIndex={-1}
+                        forceMount
+                        hidden={tab !== "Examples"}
                     >
-                        <div className="examples-list gap-2 rounded-lg">
-                            {filteredExamples().map((example, index) => (
-                                <ProjectEntry project={example} key={index} />
-                            ))}
-                        </div>
+                        {Object.entries(filteredExamples()).map((
+                            [groupName, examples],
+                            index,
+                        ) => (
+                            <Fragment key={index}>
+                                <div
+                                    className={cn(
+                                        "rounded-lg collapse collapse-arrow px-1 first:pt-0.5",
+                                        { "first:-mt-3": groupName != "all" },
+                                    )}
+                                >
+                                    <input
+                                        className="peer min-h-12"
+                                        type="checkbox"
+                                        defaultChecked
+                                        hidden={groupName == "all"}
+                                    />
+
+                                    {groupName != "all" && (
+                                        <div
+                                            className={cn(
+                                                "collapse-title flex items-center gap-1.5 font-medium pl-0 pr-6 py-3.5 min-h-12 after:!top-7 after:!right-2 peer-hover:text-white transition-colors",
+                                                {
+                                                    "text-base-content/50":
+                                                        groupName
+                                                            == "uncategorized",
+                                                },
+                                            )}
+                                        >
+                                            <span className="badge badge-xs font-bold text-[inherit] text-[0.625rem] py-1 px-1 min-w-5 h-auto bg-base-content/15 border-0">
+                                                {examples.length}
+                                            </span>
+
+                                            <div className="flex flex-wrap items-baseline gap-x-2">
+                                                <h3 className="capitalize">
+                                                    {examplesData.categories
+                                                        ?.[groupName]
+                                                        ?.displayName
+                                                        ?? groupName}
+                                                </h3>
+
+                                                {currentGroup()
+                                                        == "category"
+                                                    && examplesData
+                                                        .categories
+                                                        ?.[groupName]
+                                                        ?.description
+                                                    && (
+                                                        <p className="self-baseline font-normal text-xs tracking-wide text-base-content/80">
+                                                            {examplesData
+                                                                .categories
+                                                                ?.[groupName]
+                                                                ?.description}
+                                                        </p>
+                                                    )}
+                                            </div>
+                                        </div>
+                                    )}
+
+                                    <div
+                                        className={cn(
+                                            "examples-list gap-2 collapse-content -mx-0.5 !p-0",
+                                            {
+                                                "-mt-5 peer-checked:!py-5":
+                                                    groupName != "all",
+                                            },
+                                        )}
+                                    >
+                                        {examples.map((example) => (
+                                            <ProjectEntry
+                                                project={example}
+                                                key={example.name}
+                                                toggleTag={toggleTag}
+                                            />
+                                        ))}
+                                    </div>
+                                </div>
+                                <div className="mx-1 border-b border-base-content/10 last:hidden">
+                                </div>
+                            </Fragment>
+                        ))}
                     </Tabs.Content>
                 </Tabs.Root>
             </div>