+++ src/data/demos.ts
@@ -1,29 +1,63 @@
+import {
+    difficulties as difficultiesData,
+    tags,
+} from "../../kaplay/examples/examples.json";
 import examplesList from "./exampleList.json";
 
-export type Example = {
+export type ExamplesDataRecord = Record<string, {
+    displayName?: string;
+    description?: string;
+    order?: number;
+}>;
+
+export type Tag = {
     name: string;
-    code: string;
+} & ExamplesDataRecord[string];
+
+export type Example = {
     id: number;
-    description: string | null;
+    name: string;
     formattedName: string;
+    sortName: string;
+    category: string;
+    group: string;
+    description: string | null;
+    code: string;
     version: string;
-    tags: string[];
-    difficulty: "easy" | "medium" | "hard" | "auto" | "unknown";
+    minVersion: string;
+    tags: Tag[];
+    difficulty: {
+        level: number;
+        name: string;
+    };
+    createdAt: string;
+    updatedAt: string;
     locked?: boolean;
 };
 
 export const difficulties = [
-    "easy",
-    "medium",
-    "hard",
-    "auto",
-    "unknown",
-] as const;
+    ...difficultiesData.map(({ displayName }, index: number) => ({
+        level: index,
+        name: displayName,
+    })),
+    {
+        level: difficultiesData.length,
+        name: "Unknown",
+    },
+];
+
+export const difficultyByName = (name: string) =>
+    difficulties.find(d => d.name === name);
 
 export const demos = examplesList.map((example) => {
     const obj: Example = {
         ...example,
-        difficulty: difficulties[example.difficulty],
+        tags: example.tags.map(tag => ({
+            name: tag,
+            ...(tags as ExamplesDataRecord)?.[tag],
+        })),
+        difficulty: difficulties[example.difficulty]
+            ?? difficulties[difficulties.length - 1],
     };
 
     return obj;